geo coordunates: lat lon to decimal

+++ Abstractions/Models/GeoCoordinate.cs
@@ -7,10 +7,10 @@ namespace Filuet.Infrastructure.Abstractions.Models
     public class GeoCoordinate
     {
         [JsonPropertyName("lat")]
-        public double Latitude { get; set; }
+        public decimal Latitude { get; set; }
 
         [JsonPropertyName("lon")]
-        public double Longitude { get; set; }
+        public decimal Longitude { get; set; }
 
         public override string ToString() => string.Format("{0},{1}", Latitude, Longitude);
     }
